requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
Click==7.0
dynaconf==2.0.3
idna==2.8
iniherit==0.3.9
loguru==0.3.2
multidict==4.5.2
peewee==2.10.2
psycopg2-binary==2.7.7
python-box==3.4.2
python-dotenv==0.10.3
selectolax==0.1.13
six==1.12.0
soupsieve==1.9.2
text-unidecode==1.2
toml==0.10.0
yarl==1.3.0
yoyo-migrations==6.1.0